[#575] Move EACL Table conversion in neofs package
